Refactored getDatetime
